refactor(types): extract front types
