+++ dotcom-rendering/index.d.ts
@@ -2,6 +2,9 @@
 // CAPIArticleType and its subtypes //
 // ------------------------- //
 
+type DCRSnapType = import('./src/types/front').DCRSnapType;
+type DCRSupportingContent = import('./src/types/front').DCRSupportingContent;
+
 // Pillars are used for styling
 // RealPillars have pillar palette colours
 // FakePillars allow us to make modifications to style based on rules outside of the pillar of an article
@@ -628,413 +631,6 @@ interface CAPIArticleType {
 }
 
 type StageType = 'DEV' | 'CODE' | 'PROD';
-
-// FE* types are coming from Frontend
-interface FEFrontType {
-	pressedPage: FEPressedPageType;
-	nav: CAPINavType;
-	editionId: EditionId;
-	editionLongForm: string;
-	guardianBaseURL: string;
-	pageId: string;
-	webTitle: string;
-	webURL: string;
-	config: FEFrontConfigType;
-	commercialProperties: Record<string, unknown>;
-	pageFooter: FooterType;
-}
-
-type DCRFrontType = {
-	pressedPage: DCRPressedPageType;
-	nav: CAPINavType;
-	editionId: EditionId;
-	webTitle: string;
-	config: FEFrontConfigType;
-	pageFooter: FooterType;
-};
-
-type FEPressedPageType = {
-	id: string;
-	seoData: FESeoDataType;
-	frontProperties: FEFrontPropertiesType;
-	collections: FECollectionType[];
-};
-
-type DCRPressedPageType = {
-	id: string;
-	seoData: FESeoDataType;
-	frontProperties: FEFrontPropertiesType;
-	collections: DCRCollectionType[];
-};
-
-type FESeoDataType = {
-	id: string;
-	navSection: string;
-	webTitle: string;
-	description: string;
-};
-
-type FEFrontPropertiesType = {
-	isImageDisplayed: boolean;
-	commercial: Record<string, unknown>;
-};
-
-type FESupportingContent = {
-	properties: {
-		href?: string;
-	};
-	header?: {
-		kicker?: {
-			item?: {
-				properties: {
-					kickerText: string;
-				};
-			};
-		};
-		headline: string;
-		url: string;
-	};
-	format?: CAPIFormat;
-};
-
-type DCRSupportingContent = {
-	headline: string;
-	url?: string;
-	kickerText?: string;
-	format: ArticleFormat;
-};
-
-type FESnapType = {
-	embedHtml?: string;
-	embedCss?: string;
-	embedJs?: string;
-};
-
-type DCRSnapType = {
-	embedHtml?: string;
-	embedCss?: string;
-};
-
-type FEContainerType =
-	| 'dynamic/fast'
-	| 'dynamic/package'
-	| 'dynamic/slow'
-	| 'dynamic/slow-mpu'
-	| 'fixed/large/slow-XIV'
-	| 'fixed/medium/fast-XI'
-	| 'fixed/medium/fast-XII'
-	| 'fixed/medium/slow-VI'
-	| 'fixed/medium/slow-VII'
-	| 'fixed/medium/slow-XII-mpu'
-	| 'fixed/small/fast-VIII'
-	| 'fixed/small/slow-I'
-	| 'fixed/small/slow-III'
-	| 'fixed/small/slow-IV'
-	| 'fixed/small/slow-V-half'
-	| 'fixed/small/slow-V-mpu'
-	| 'fixed/small/slow-V-third'
-	| 'fixed/thrasher'
-	| 'fixed/video'
-	| 'nav/list'
-	| 'nav/media-list'
-	| 'news/most-popular';
-
-type FEContainerPalette =
-	| 'EventPalette'
-	| 'SombreAltPalette'
-	| 'EventAltPalette'
-	| 'InvestigationPalette'
-	| 'LongRunningAltPalette'
-	| 'LongRunningPalette'
-	| 'SombrePalette'
-	| 'Canonical'
-	| 'Dynamo'
-	| 'Special'
-	| 'DynamoLike'
-	| 'Special'
-	| 'Breaking'
-	| 'Podcast'
-	| 'Branded'
-	| 'BreakingPalette';
-
-type DCRContainerPalette =
-	| 'EventPalette'
-	| 'SombreAltPalette'
-	| 'EventAltPalette'
-	| 'InvestigationPalette'
-	| 'LongRunningAltPalette'
-	| 'LongRunningPalette'
-	| 'SombrePalette'
-	| 'BreakingPalette';
-
-// TODO: These may need to be declared differently than the front types in the future
-type DCRContainerType = FEContainerType;
-
-type FEFrontCard = {
-	properties: {
-		isBreaking: boolean;
-		showMainVideo: boolean;
-		showKickerTag: boolean;
-		showByline: boolean;
-		imageSlideshowReplace: boolean;
-		maybeContent?: {
-			trail: {
-				trailPicture?: {
-					allImages: {
-						index: number;
-						fields: {
-							displayCredit?: string;
-							source?: string;
-							photographer?: string;
-							isMaster?: string;
-							altText?: string;
-							height: string;
-							credit: string;
-							mediaId: string;
-							width: string;
-						};
-						mediaType: string;
-						url: string;
-					}[];
-				};
-				byline?: string;
-				thumbnailPath?: string;
-				webPublicationDate: number;
-			};
-			metadata: {
-				id: string;
-				webTitle: string;
-				webUrl: string;
-				type: string;
-				sectionId?: { value: string };
-				format: CAPIFormat;
-			};
-			fields: {
-				main: string;
-				body: string;
-				standfirst?: string;
-			};
-			elements: Record<string, unknown>;
-			tags: Record<string, unknown>;
-		};
-		maybeContentId?: string;
-		isLiveBlog: boolean;
-		isCrossword: boolean;
-		byline?: string;
-		webTitle: string;
-		linkText?: string;
-		webUrl?: string;
-		editionBrandings: { edition: { id: EditionId } }[];
-	};
-	header: {
-		isVideo: boolean;
-		isComment: boolean;
-		isGallery: boolean;
-		isAudio: boolean;
-		kicker?: {
-			type: string;
-			item?: {
-				properties: {
-					kickerText: string;
-				};
-			};
-		};
-		seriesOrBlogKicker?: {
-			properties: {
-				kickerText: string;
-			};
-			name: string;
-			url: string;
-			id: string;
-		};
-		headline: string;
-		url: string;
-		hasMainVideoElement: boolean;
-	};
-	card: {
-		id: string;
-		cardStyle: {
-			type: string;
-		};
-		webPublicationDateOption?: number;
-		lastModifiedOption?: number;
-		trailText?: string;
-		starRating?: number;
-		shortUrlPath?: string;
-		shortUrl: string;
-		group: string;
-		isLive: boolean;
-	};
-	discussion: {
-		isCommentable: boolean;
-		isClosedForComments: boolean;
-		discussionId?: string;
-	};
-	display: {
-		isBoosted: boolean;
-		showBoostedHeadline: boolean;
-		showQuotedHeadline: boolean;
-		imageHide: boolean;
-		showLivePlayable: boolean;
-	};
-	format?: CAPIFormat;
-	enriched?: FESnapType;
-	supportingContent?: FESupportingContent[];
-	cardStyle?: {
-		type: string;
-	};
-	type: string;
-};
-
-type DCRFrontCard = {
-	format: ArticleFormat;
-	url: string;
-	headline: string;
-	trailText?: string;
-	webPublicationDate?: string;
-	image?: string;
-	kickerText?: string;
-	snapData?: DCRSnapType;
-	/** @see JSX.IntrinsicAttributes["data-link-name"] */
-	dataLinkName: string;
-	discussionId?: string;
-	byline?: string;
-	showByline?: boolean;
-	isBoosted?: boolean;
-};
-
-type FECollectionType = {
-	id: string;
-	displayName: string;
-	curated: FEFrontCard[];
-	backfill: FEFrontCard[];
-	treats: FEFrontCard[];
-	lastUpdate?: number;
-	href?: string;
-	groups?: string[];
-	collectionType: FEContainerType;
-	uneditable: boolean;
-	showTags: boolean;
-	showSections: boolean;
-	hideKickers: boolean;
-	showDateHeader: boolean;
-	showLatestUpdate: boolean;
-	config: {
-		displayName: string;
-		metadata?: { type: FEContainerPalette }[];
-		collectionType: FEContainerType;
-		href?: string;
-		groups?: string[];
-		uneditable: boolean;
-		showTags: boolean;
-		showSections: boolean;
-		hideKickers: boolean;
-		showDateHeader: boolean;
-		showLatestUpdate: boolean;
-		excludeFromRss: boolean;
-		showTimestamps: boolean;
-		hideShowMore: boolean;
-		platform: string;
-	};
-	hasMore: boolean;
-};
-
-type DCRCollectionType = {
-	id: string;
-	displayName: string;
-	collectionType: DCRContainerType;
-	containerPalette?: DCRContainerPalette;
-	curated: DCRFrontCard[];
-	backfill: DCRFrontCard[];
-	treats: DCRFrontCard[];
-	href?: string;
-	config: {
-		showDateHeader: boolean;
-	};
-};
-
-type FEFrontConfigType = {
-	avatarApiUrl: string;
-	externalEmbedHost: string;
-	ajaxUrl: string;
-	keywords: string;
-	revisionNumber: string;
-	isProd: boolean;
-	switches: Switches;
-	section: string;
-	keywordIds: string;
-	locationapiurl: string;
-	sharedAdTargeting: { [key: string]: any };
-	buildNumber: string;
-	abTests: ServerSideTests;
-	pbIndexSites: { [key: string]: unknown }[];
-	ampIframeUrl: string;
-	beaconUrl: string;
-	userAttributesApiUrl: string;
-	host: string;
-	brazeApiKey: string;
-	calloutsUrl: string;
-	requiresMembershipAccess: boolean;
-	onwardWebSocket: string;
-	a9PublisherId: string;
-	contentType: string;
-	facebookIaAdUnitRoot: string;
-	ophanEmbedJsUrl: string;
-	idUrl: string;
-	dcrSentryDsn: string;
-	isFront: true;
-	idWebAppUrl: string;
-	discussionApiUrl: string;
-	sentryPublicApiKey: string;
-	omnitureAccount: string;
-	dfpAccountId: string;
-	pageId: string;
-	forecastsapiurl: string;
-	assetsPath: string;
-	pillar: string;
-	commercialBundleUrl: string;
-	discussionApiClientHeader: string;
-	membershipUrl: string;
-	dfpHost: string;
-	cardStyle?: string;
-	googletagUrl: string;
-	sentryHost: string;
-	shouldHideAdverts: boolean;
-	mmaUrl: string;
-	membershipAccess: string;
-	isPreview: boolean;
-	googletagJsUrl: string;
-	supportUrl: string;
-	edition: string;
-	discussionFrontendUrl: string;
-	ipsosTag: string;
-	ophanJsUrl: string;
-	isPaidContent: boolean;
-	mobileAppsAdUnitRoot: string;
-	plistaPublicApiKey: string;
-	frontendAssetsFullURL: string;
-	googleSearchId: string;
-	allowUserGeneratedContent: boolean;
-	dfpAdUnitRoot: string;
-	idApiUrl: string;
-	omnitureAmpAccount: string;
-	adUnit: string;
-	hasPageSkin: boolean;
-	webTitle: string;
-	stripePublicToken: string;
-	googleRecaptchaSiteKey: string;
-	discussionD2Uid: string;
-	weatherapiurl: string;
-	googleSearchUrl: string;
-	optimizeEpicUrl: string;
-	stage: StageType;
-	idOAuthUrl: string;
-	isSensitive: boolean;
-	isDev: boolean;
-	thirdPartyAppsAccount?: string;
-	avatarImagesUrl: string;
-	fbAppId: string;
-};
 
 interface TagType {
 	id: string;
