+++ dotcom-rendering/src/model/enhanceCollections.ts
@@ -1,3 +1,4 @@
+import type { DCRCollectionType, FECollectionType } from '../types/front';
 import { decideContainerPalette } from './decideContainerPalette';
 import { enhanceCards } from './enhanceCards';
