+++ dotcom-rendering/src/model/enhanceSnaps.ts
@@ -1,6 +1,8 @@
+import type { DCRSnapType, FESnapType } from '../types/front';
+
 /**
  *
- * This function takes a FESnapType and turns it into a DCRSnapType. It mutates
+ * This function takes a `FESnapType` and turns it into a `DCRSnapType`. It mutates
  * certain properties such as embedCss to work with Emotion.
  *
  * @returns the DCR snap with its properties mutated
@@ -11,7 +13,7 @@ export const enhanceSnaps = (
 	const dcrSnap = snap;
 
 	// Emotion doesn't parse this conditional CSS correctly which is used by some thrashers to check
-	// if theres a pageskin. This regex fixes the vast majority of these cases should allow future snaps
+	// if there’s a page skin. This regex fixes the vast majority of these cases should allow future snaps
 	// to be built with DCR in mind.
 	if (dcrSnap?.embedCss) {
 		dcrSnap.embedCss = dcrSnap.embedCss.replace(
