+++ dotcom-rendering/src/model/window-guardian.ts
@@ -1,3 +1,5 @@
+import type { DCRFrontType } from '../types/front';
+
 export interface WindowGuardianConfig {
 	isDotcomRendering: boolean;
 	isDev: boolean;
