+++ dotcom-rendering/src/web/components/Card/components/CardAge.tsx
@@ -2,6 +2,7 @@ import { css } from '@emotion/react';
 import { ArticleDesign, timeAgo } from '@guardian/libs';
 import { textSans, until } from '@guardian/source-foundations';
 import ClockIcon from '../../../../static/icons/clock.svg';
+import type { DCRContainerPalette } from '../../../../types/front';
 import { decidePalette } from '../../../lib/decidePalette';
 
 type Props = {
