+++ dotcom-rendering/src/web/components/Card/components/TrailTextWrapper.tsx
@@ -1,5 +1,6 @@
 import { css } from '@emotion/react';
 import { body, until } from '@guardian/source-foundations';
+import type { DCRContainerPalette } from '../../../../types/front';
 import { decidePalette } from '../../../lib/decidePalette';
 
 type Props = {
