+++ dotcom-rendering/src/web/components/ContainerTitle.tsx
@@ -8,6 +8,7 @@ import {
 	text,
 	until,
 } from '@guardian/source-foundations';
+import type { DCRContainerPalette } from '../../types/front';
 import { decideContainerOverrides } from '../lib/decideContainerOverrides';
 import { getEditionFromId } from '../lib/edition';
