+++ dotcom-rendering/src/web/components/DynamicFast.tsx
@@ -1,3 +1,4 @@
+import type { DCRContainerPalette } from '../../types/front';
 import { LI } from './Card/components/LI';
 import { UL } from './Card/components/UL';
 import { FrontCard } from './FrontCard';
