+++ dotcom-rendering/src/web/components/MediaMeta.tsx
@@ -5,6 +5,7 @@ import {
 	SvgCamera,
 	SvgVideo,
 } from '@guardian/source-react-components';
+import type { DCRContainerPalette } from '../../types/front';
 import { decidePalette } from '../lib/decidePalette';
 
 type Props = {
