+++ dotcom-rendering/src/web/components/Snap.tsx
@@ -1,4 +1,5 @@
 import { css } from '@emotion/react';
+import type { DCRSnapType } from '../../types/front';
 
 const snapStyles = css`
 	overflow-y: hidden;
