+++ dotcom-rendering/src/web/components/Spotlight.tsx
@@ -1,4 +1,5 @@
 import { ArticleDesign } from '@guardian/libs';
+import type { DCRContainerPalette } from '../../types/front';
 import { Card } from './Card/Card';
 
 type Props = {
