+++ dotcom-rendering/src/web/components/SupportingContent.tsx
@@ -1,5 +1,6 @@
 import { css } from '@emotion/react';
 import { from, until } from '@guardian/source-foundations';
+import type { DCRSupportingContent } from '../../types/front';
 import { CardHeadline } from './CardHeadline';
 
 type Alignment = 'vertical' | 'horizontal';
